[Serialize] Fix a bug that when template is a function, htmlAttribute doesn't take effect.
The root cause of this bug is that in serailize.js, when template
is a function, widget is generated without applying html
attributes. To simplify the logic, I refactor the getTemplate
fucntion so that it only retures string or object. Note that in
palette.js, the "code" data is not used by any other code so I
delete the line setting the "code" data.
Note: image-tag is dependent on this submission.